requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
diffusers
transformers
accelerate
pillow
onnxruntime
optimum
safetensors
numpy
huggingface-hub
Flask
